refactor: split safe action into server and client
