+++ components/groups/endpoints/options-dropdown.tsx
@@ -28,7 +28,7 @@ import {
 } from "@/lib/data/endpoints";
 
 import { useAction } from "next-safe-action/hooks";
-import { parseActionError } from "@/lib/data/safe-action";
+import { parseActionError } from "@/lib/data/safe-action.client";
 import { toast } from "sonner";
 
 export default function OptionsDropdown({
