+++ components/groups/leads/options-dropdown.tsx
@@ -22,7 +22,7 @@ import { useState } from "react";
 
 import { deleteLead } from "@/lib/data/leads";
 import { useAction } from "next-safe-action/hooks";
-import { parseActionError } from "@/lib/data/safe-action";
+import { parseActionError } from "@/lib/data/safe-action.client";
 import { toast } from "sonner";
 
 export default function OptionsDropdown({ id }: { id: string }) {
